updated parallelizer names to make sense
