+++ bond_analysis.py
@@ -12,7 +12,7 @@ import numpy as np
 from sys import exit, stderr
 from output_bonds import output_bonds
 import argparse
-from UTILS import parallelize
+from UTILS import parallelize_lorenzo_onefile
 from os import environ
 
 def bond_analysis(reader, pairs, num_confs, start=None, stop=None):
@@ -132,7 +132,7 @@ if __name__ == "__main__":
 
     if parallel:
         print("INFO: Computing base pairs in {} configurations using {} cores.".format(num_confs, n_cpus), file=stderr)
-        out = parallelize.fire_multiprocess(traj_file, top_file, bond_analysis, num_confs, n_cpus, pairs)
+        out = parallelize_lorenzo_onefile.fire_multiprocess(traj_file, top_file, bond_analysis, num_confs, n_cpus, pairs)
         tot_bonds = sum((i[0] for i in out))
         tot_missbonds = sum((i[1] for i in out))
         out_array = sum((i[2] for i in out))
