+++ tests/cleanup.sh
@@ -1,1 +1,1 @@
-rm aligned.dat all_energy.json angles.txt angle.png animated.mp4 centroid.dat cluster_0.dat cluster_data.json coordinates.png devs.json devsM.json distance.png mean.dat meanM.dat meanM.top pairs.json pca.json scree.png
+rm aligned.dat all_energy.json angles.txt angle.png animated.mp4 centroid.dat cluster_0.dat cluster_data.json coordinates.png devs.json devsM.json distance.png mean.dat meanM.dat meanM.top pairs.json pca* scree.png
